+++ sdk/communication/azure-communication-identity/src/main/java/com/azure/communication/identity/package-info.java
@@ -1,5 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-
-/** Package containing the classes for AzureCommunicationIdentity. Azure Communication Identity Service. */
+/**
+ * Package containing the classes for AzureCommunicationIdentity. Azure Communication Identity Service.
+ */
 package com.azure.communication.identity;